# server_safl/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "South African Fishing Log Application API",
    "description": "REST API Documentation",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3443",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID of user that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user with given ID",
        "responses": {
          "200": {
            "description": "User is found",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users/register": {
      "parameters": [
        {
          "name": "user",
          "in": "body",
          "description": "User that we want to register",
          "schema": {
            "$ref": "#/definitions/User"
          }
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Register a new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to register",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is registered",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users/login": {
      "parameters": [
        {
          "name": "user",
          "in": "body",
          "description": "User that we want to login",
          "schema": {
            "$ref": "#/definitions/User"
          }
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "description": "User Login",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User Login",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User is logged in",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "required": [
        "email","username","password","admin",
        "_id"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "admin": {
          "type": "boolean"
        }
      }
    },
...
  }
}